webserver/requirements.txt
click==8.0.3
Flask==2.0.2
gevent==21.8.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.1
grpcio-tools==1.41.1
gunicorn==20.1.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pkg_resources==0.0.0
protobuf==3.19.1
six==1.16.0
typing_extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
